✨ feat: 2-1 API 연결 함수 구현
