+++ src/type/formtype.d.ts
@@ -63,38 +63,6 @@ declare interface ClubCardProps {
   isLarge?: boolean; //크기 변환 prop
 }
 
-// 합격 인원 설정 Form
-declare interface SetAcceptanceCountFormData {
-  totalDocumentPassCount: number;
-  totalFinalPassCount: number;
-  groupInfos: {
-    groupName: string;
-    documentPassCount: number;
-    finalPassCount: number;
-  }[];
-}
-
-declare interface GroupPassCardProps {
-  control: Control<any>;
-  groupIndex: number;
-  groupName: string;
-  errors: any;
-  rules: {
-    groupName?: any;
-    documentPassCount?: any;
-    finalPassCount?: any;
-  };
-}
-
-declare interface GroupPassCountProps {
-  control: Control<any>;
-  errors: any;
-  rules: {
-    groupName?: any;
-    documentPassCount?: any;
-    finalPassCount?: any;
-  };
-}
 // 운영자 Type
 declare interface AdminUser {
   id: string;
